RF: Replace BIDSLayout with BIDSLayoutV2
