+++ bids/layout/__init__.py
@@ -1,5 +1,3 @@
-import warnings
-
 from .layout import BIDSLayout, Query
 from .models import (BIDSFile, BIDSImageFile, BIDSDataFile, BIDSJSONFile,
                      Config, Entity, Tag)
@@ -22,11 +20,4 @@ __all__ = [
     "Entity",
     "Tag",
     "Query",
-    "BIDSLayoutV2",
 ]
-
-try:
-    from .layout_v2 import BIDSLayoutV2
-except Exception as err:
-    def BIDSLayoutV2(*args, **kwargs):
-        raise RuntimeError("Cannot create BIDSLayoutV2 - please install the ancpbids package.") from err
